Move TimestampOperations to utils package
