+++ backend/src/main/java/com/odde/doughnut/entities/MemoryTracker.java
@@ -1,7 +1,7 @@
 package com.odde.doughnut.entities;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.odde.doughnut.models.TimestampOperations;
+import com.odde.doughnut.utils.TimestampOperations;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
