+++ backend/src/main/java/com/odde/doughnut/services/AssimilationService.java
@@ -6,8 +6,8 @@ import com.odde.doughnut.entities.Note;
 import com.odde.doughnut.factoryServices.ModelFactoryService;
 import com.odde.doughnut.models.ReviewScope;
 import com.odde.doughnut.models.SubscriptionModel;
-import com.odde.doughnut.models.TimestampOperations;
 import com.odde.doughnut.models.UserModel;
+import com.odde.doughnut.utils.TimestampOperations;
 import java.sql.Timestamp;
 import java.time.ZoneId;
 import java.util.List;
