+++ backend/src/main/java/com/odde/doughnut/services/wikidataApis/WikidataValue.java
@@ -2,8 +2,8 @@ package com.odde.doughnut.services.wikidataApis;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.odde.doughnut.entities.Coordinate;
-import com.odde.doughnut.models.TimestampOperations;
 import com.odde.doughnut.services.wikidataApis.thirdPartyEntities.WikidataDatavalue;
+import com.odde.doughnut.utils.TimestampOperations;
 
 public record WikidataValue(WikidataDatavalue datavalue) {
