+++ backend/src/main/java/com/odde/doughnut/testability/TestabilityRestController.java
@@ -7,12 +7,12 @@ import com.odde.doughnut.entities.*;
 import com.odde.doughnut.entities.repositories.NoteRepository;
 import com.odde.doughnut.entities.repositories.UserRepository;
 import com.odde.doughnut.factoryServices.ModelFactoryService;
-import com.odde.doughnut.models.TimestampOperations;
 import com.odde.doughnut.models.UserModel;
 import com.odde.doughnut.services.CircleService;
 import com.odde.doughnut.services.GithubService;
 import com.odde.doughnut.services.NoteConstructionService;
 import com.odde.doughnut.testability.model.PredefinedQuestionsTestData;
+import com.odde.doughnut.utils.TimestampOperations;
 import jakarta.persistence.EntityManagerFactory;
 import java.io.IOException;
 import java.sql.Timestamp;
