+++ backend/src/main/java/com/odde/doughnut/utils/TimestampOperations.java
@@ -1,4 +1,4 @@
-package com.odde.doughnut.models;
+package com.odde.doughnut.utils;
 
 import java.sql.Timestamp;
 import java.time.Instant;
