+++ backend/src/test/java/com/odde/doughnut/models/RecallServiceWithSpacedRepetitionAlgorithmTest.java
@@ -8,6 +8,7 @@ import static org.hamcrest.Matchers.nullValue;
 import com.odde.doughnut.entities.MemoryTracker;
 import com.odde.doughnut.entities.Note;
 import com.odde.doughnut.testability.MakeMe;
+import com.odde.doughnut.utils.TimestampOperations;
 import java.sql.Timestamp;
 import java.time.ZoneId;
 import org.junit.jupiter.api.BeforeEach;
