+++ backend/src/test/java/com/odde/doughnut/models/TimestampOperationsTest.java
@@ -3,6 +3,7 @@ package com.odde.doughnut.models;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 
+import com.odde.doughnut.utils.TimestampOperations;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
